omra_lite/backend/requirements.txt
# Core dependencies
fastapi==0.95.1
uvicorn==0.22.0
pydantic==2.0.3
python-dotenv==1.0.0
motor==3.1.2
pymongo==4.3.3

# Security and authentication
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.0.0.post2

# HTTP and API
httpx==0.24.1
aiofiles==23.1.0

# AI integration (optional - uncomment if needed)
 anthropic==0.2.8
 openai==0.27.6

# Utilities
jinja2==3.1.2
python-dateutil==2.8.2
pytest==7.3.1
bcrypt==4.0.1
pycryptodome==3.17

# Additional dependencies
python-jose[cryptography]==3.3.0 